Add run() to thread and add start() to main.
The run() method sets a random time length between 0 and 10 seconds then repeatedly prints out ( name var + "is eating") of the object that is doing the run(). start() in Main starts the run method for each object. I'm not sure if the objects pre-empt each other or wait for the current eater to finish, but they do "eat" for different lengths of time because the timer is set to a random length. We now need to add: 1. the hungry and full conditions 2. the chopsticks counter and condition to only take when counter > 2