Try approach to chopsticks counter
Experiment with idea: int chopsticks doesn't need to be part of thread. chopsticks is just a counter, and the run method takes 2 chopsticks when it starts and puts back 2 when it finishes. But where should chopsticks be declared and initialized? As written, the code starts by having all five objects eating at the same time.