Adds missing } and inserts comments
This version will compile and run in Java. Still missing: 1. how to increment the counter 2. how to prevent the neighbors from eating when the philosopher between them is eating 3. how to have the eater(s) get full and stop eating 4. how to get the ones have been full hungry again after a while.