Run() issues resolved
fixed the problem of run() not recognizing the array by declaring the array and all of the philosopher objects INSIDE the Philosopher class. fixed the problem of the start() methods in main  not recognizing the philosoper objects by making them static.
Wrote check left and check right methods and call them in run() which seems to be working.